+++ CHANGELOG.md
@@ -11,3 +11,10 @@ The format is based on [Keep a Changelog] and this project adheres to
 ### Added
 - Changelog, license, readme.
 - Gradle JOSM plugin demo.
+- Batch Circle Building action.
+- Batch Orthogonal Building action.
+- Pick Residential Area action.
+- Mapathoner main class.
+
+### Changed
+- Update `build.gradle` with actual project info.